@@ -1,5 +1,5 @@
 {
-  "version": "5.0.0",
+  "version": "5.0.1",
   "private": true,
   "scripts": {
     "start": "yarn docs:dev",
@@ -55,18 +55,18 @@
     "@babel/core": "^7.16.0",
     "@babel/node": "^7.16.0",
     "@babel/plugin-transform-object-assign": "^7.16.0",
-    "@babel/plugin-transform-runtime": "^7.16.0",
+    "@babel/plugin-transform-runtime": "^7.16.4",
     "@babel/register": "^7.16.0",
     "@babel/runtime": "^7.16.3",
-    "@emotion/cache": "^11.5.0",
-    "@emotion/react": "^11.5.0",
-    "@emotion/styled": "^11.3.0",
+    "@emotion/cache": "^11.6.0",
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
     "@eps1lon/enzyme-adapter-react-17": "^0.1.0",
     "@material-ui/monorepo": "https://github.com/mui-org/material-ui.git#master",
-    "@mui/icons-material": "^5.1.0",
-    "@mui/material": "^5.1.0",
-    "@mui/styles": "^5.1.0",
-    "@mui/utils": "^5.1.0",
+    "@mui/icons-material": "^5.1.1",
+    "@mui/material": "^5.1.1",
+    "@mui/styles": "^5.1.1",
+    "@mui/utils": "^5.1.1",
     "@rollup/plugin-babel": "^5.3.0",
     "@rollup/plugin-node-resolve": "^13.0.6",
     "@rollup/plugin-replace": "^3.0.0",
@@ -77,13 +77,13 @@
     "@types/mocha": "^9.0.0",
     "@types/node": "^16.11.4",
     "@types/prettier": "^2.4.2",
-    "@types/react": "^17.0.30",
+    "@types/react": "^17.0.36",
     "@types/react-dom": "^17.0.11",
     "@types/requestidlecallback": "^0.3.4",
     "@types/sinon": "^10.0.6",
-    "@types/yargs": "^17.0.5",
-    "@typescript-eslint/eslint-plugin": "^5.3.1",
-    "@typescript-eslint/parser": "^5.3.1",
+    "@types/yargs": "^17.0.7",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
     "argos-cli": "^0.3.3",
     "babel-plugin-module-resolver": "^4.1.0",
     "babel-plugin-optimize-clsx": "^2.6.2",
@@ -91,7 +91,7 @@
     "babel-plugin-transform-rename-import": "^2.3.0",
     "chai": "^4.3.4",
     "chai-dom": "^1.10.0",
-    "compression-webpack-plugin": "^9.0.0",
+    "compression-webpack-plugin": "^9.0.1",
     "concurrently": "^6.4.0",
     "cross-env": "^7.0.3",
     "enzyme": "^3.11.0",
@@ -104,7 +104,7 @@
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-mocha": "^9.0.0",
     "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-react": "^7.27.0",
+    "eslint-plugin-react": "^7.27.1",
     "eslint-plugin-react-hooks": "^4.3.0",
     "format-util": "^1.0.5",
     "glob-gitignore": "^1.0.14",
@@ -112,7 +112,7 @@
     "jss": "^10.8.2",
     "jss-plugin-template": "^10.8.2",
     "jss-rtl": "^0.3.0",
-    "karma": "^6.3.8",
+    "karma": "^6.3.9",
     "karma-browserstack-launcher": "^1.6.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^2.0.1",
@@ -133,15 +133,15 @@
     "rollup-plugin-dts": "^3.0.2",
     "rollup-plugin-sourcemaps": "^0.6.3",
     "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.30.0",
+    "rollup-plugin-typescript2": "^0.31.0",
     "serve": "^13.0.2",
     "sinon": "^12.0.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^22.0.0",
     "stylelint-processor-styled-components": "^1.10.0",
-    "typedoc": "^0.22.8",
-    "typescript": "^4.4.4",
-    "webpack": "^5.64.0",
+    "typedoc": "^0.22.9",
+    "typescript": "^4.5.2",
+    "webpack": "^5.64.2",
     "webpack-cli": "^4.9.1",
     "yargs": "^17.2.1",
     "yarn-deduplicate": "^3.1.0"
