@@ -21,7 +21,16 @@ const useStyles = makeStyles({
 
 function RatingInputValue(props: GridFilterInputValueProps) {
   const classes = useStyles();
-  const { item, applyValue } = props;
+  const { item, applyValue, focusElementRef } = props;
+
+  const ratingRef: React.Ref<any> = React.useRef(null);
+  React.useImperativeHandle(focusElementRef, () => ({
+    focus: () => {
+      ratingRef.current
+        .querySelector(`input[value="${Number(item.value) || ''}"]`)
+        .focus();
+    },
+  }));
 
   const handleFilterChange = (event) => {
     applyValue({ ...item, value: event.target.value });
@@ -35,6 +44,7 @@ function RatingInputValue(props: GridFilterInputValueProps) {
         value={Number(item.value)}
         onChange={handleFilterChange}
         precision={0.5}
+        ref={ratingRef}
       />
     </div>
   );
@@ -45,7 +55,7 @@ export default function ExtendNumericOperator() {
   const columns = [...data.columns];
 
   const [filterModel, setFilterModel] = React.useState<GridFilterModel>({
-    items: [{ columnField: 'rating', value: '3.5', operatorValue: '>=' }],
+    items: [{ id: 1, columnField: 'rating', value: '3.5', operatorValue: '>=' }],
   });
 
   if (columns.length > 0) {
