@@ -1,6 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { unstable_composeClasses as composeClasses } from '@mui/core';
+import { unstable_composeClasses as composeClasses } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
 import Box from '@mui/material/Box';
 import { useGridRootProps } from '../../hooks/utils/useGridRootProps';
@@ -53,7 +53,7 @@ const GridTreeDataGroupingCell = (props: GridRenderCellParams<GridTreeDataGroupi
   };
 
   const handleClick = (event) => {
-    apiRef.current.unstable_setRowExpansion(id, !value.expanded);
+    apiRef.current.setRowChildrenExpansion(id, !value.expanded);
     apiRef.current.setCellFocus(id, field);
     event.stopPropagation();
   };
@@ -145,8 +145,8 @@ GridTreeDataGroupingCell.propTypes = {
     children: PropTypes.arrayOf(
       PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
     ),
+    childrenExpanded: PropTypes.bool,
     depth: PropTypes.number.isRequired,
-    expanded: PropTypes.bool,
     groupingValue: PropTypes.string.isRequired,
     id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
     isAutoGenerated: PropTypes.bool,
