@@ -141,8 +141,8 @@ GridCellCheckboxForwardRef.propTypes = {
     children: PropTypes.arrayOf(
       PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
     ),
+    childrenExpanded: PropTypes.bool,
     depth: PropTypes.number.isRequired,
-    expanded: PropTypes.bool,
     groupingValue: PropTypes.string.isRequired,
     id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
     isAutoGenerated: PropTypes.bool,
