@@ -56,9 +56,9 @@ const deDEGrid: Partial<GridLocaleText> = {
   filterOperatorEndsWith: 'endet mit',
   filterOperatorIs: 'ist',
   filterOperatorNot: 'ist nicht',
-  filterOperatorOnOrAfter: 'ist an oder nach',
+  filterOperatorOnOrAfter: 'ist am oder nach',
   filterOperatorBefore: 'ist vor',
-  filterOperatorOnOrBefore: 'ist an oder vor',
+  filterOperatorOnOrBefore: 'ist am oder vor',
   filterOperatorAfter: 'ist nach',
   filterOperatorIsEmpty: 'ist leer',
   filterOperatorIsNotEmpty: 'ist nicht leer',
@@ -92,7 +92,7 @@ const deDEGrid: Partial<GridLocaleText> = {
     `${visibleCount.toLocaleString()} von ${totalCount.toLocaleString()}`,
 
   // Actions cell more text
-  // actionsCellMore: 'more',
+  actionsCellMore: 'Mehr',
 };
 
 export const deDE: Localization = getGridLocalization(deDEGrid, deDECore);
