@@ -4,7 +4,7 @@ import { getGridLocalization, Localization } from '../utils/getGridLocalization'
 const elGRGrid: Partial<GridLocaleText> = {
   // Root
   noRowsLabel: 'Δεν υπάρχουν καταχωρήσεις',
-  // noResultsOverlayLabel: 'No results found.',
+  noResultsOverlayLabel: 'Δεν βρέθηκαν αποτελέσματα.',
   errorOverlayDefaultLabel: 'Παρουσιάστηκε απρόβλεπτο σφάλμα.',
 
   // Density selector toolbar button text
@@ -59,8 +59,8 @@ const elGRGrid: Partial<GridLocaleText> = {
   filterOperatorOnOrAfter: 'είναι ίσο ή μετά',
   filterOperatorBefore: 'είναι πριν',
   filterOperatorOnOrBefore: 'είναι ίσο ή πριν',
-  // filterOperatorIsEmpty: 'is empty',
-  // filterOperatorIsNotEmpty: 'is not empty',
+  filterOperatorIsEmpty: 'είναι κενό',
+  filterOperatorIsNotEmpty: 'δεν είναι κενό',
 
   // Column menu text
   columnMenuLabel: 'Μενού',
@@ -87,11 +87,11 @@ const elGRGrid: Partial<GridLocaleText> = {
   footerTotalRows: 'Σύνολο Γραμμών:',
 
   // Total visible rows footer text
-  // footerTotalVisibleRows: (visibleCount, totalCount) =>
-  //   `${visibleCount.toLocaleString()} of ${totalCount.toLocaleString()}`,
+  footerTotalVisibleRows: (visibleCount, totalCount) =>
+    `${visibleCount.toLocaleString()} από ${totalCount.toLocaleString()}`,
 
   // Actions cell more text
-  // actionsCellMore: 'more',
+  actionsCellMore: 'περισσότερα',
 };
 
 export const elGR: Localization = getGridLocalization(elGRGrid);
