@@ -39,14 +39,12 @@ export interface GridRowApi {
    * Gets the row node from the internal tree structure.
    * @param {GridRowId} id The id of the row.
    * @returns {GridRowTreeNodeConfig} The row data.
-   * @ignore - do not document.
    */
-  unstable_getRowNode: (id: GridRowId) => GridRowTreeNodeConfig | null;
+  getRowNode: (id: GridRowId) => GridRowTreeNodeConfig | null;
   /**
    * Expand or collapse a row children.
    * @param {GridRowId} id the ID of the row to expand or collapse.
    * @param {boolean} isExpanded A boolean indicating if the row must be expanded or collapsed.
-   * @ignore - do not document.
    */
-  unstable_setRowExpansion: (id: GridRowId, isExpanded: boolean) => void;
+  setRowChildrenExpansion: (id: GridRowId, isExpanded: boolean) => void;
 }
