@@ -243,23 +243,19 @@ export interface GridColumnsMeta {
   positions: number[];
 }
 
-export type GridColumnLookup = { [field: string]: GridStateColDef };
-
-export type GridRawColumnLookup = { [field: string]: GridStateColDef | GridColDef };
-
-export interface GridColumnsState {
-  all: string[];
-  lookup: GridColumnLookup;
-}
+export type GridColDefOverride<ForcedFields extends keyof GridColDef> = Omit<
+  Partial<GridColDef>,
+  ForcedFields
+>;
 
 export interface GridRawColumnsState {
   all: string[];
   lookup: GridRawColumnLookup;
 }
 
-export type GridColDefOverride = Omit<Partial<GridColDef>, 'field'>;
-
-export type GridColDefOverrideCallback = (params: GridColDefOverrideParams) => GridColDefOverride;
+export type GridColDefOverrideCallback<ForcedFields extends keyof GridColDef> = (
+  params: GridColDefOverrideParams,
+) => GridColDefOverride<ForcedFields>;
 
 export interface GridColDefOverrideParams {
   /**
