@@ -19,11 +19,14 @@ describe('buildRowTree', () => {
     });
 
     expect(
-      Object.values(response.tree).map((node) => ({ id: node.id, expanded: node.expanded })),
+      Object.values(response.tree).map((node) => ({
+        id: node.id,
+        childrenExpanded: node.childrenExpanded,
+      })),
     ).to.deep.equal([
-      { id: 0, expanded: false },
-      { id: 1, expanded: false },
-      { id: 2, expanded: false },
+      { id: 0, childrenExpanded: false },
+      { id: 1, childrenExpanded: false },
+      { id: 2, childrenExpanded: false },
     ]);
   });
 
@@ -44,11 +47,14 @@ describe('buildRowTree', () => {
     });
 
     expect(
-      Object.values(response.tree).map((node) => ({ id: node.id, expanded: node.expanded })),
+      Object.values(response.tree).map((node) => ({
+        id: node.id,
+        childrenExpanded: node.childrenExpanded,
+      })),
     ).to.deep.equal([
-      { id: 0, expanded: true },
-      { id: 1, expanded: true },
-      { id: 2, expanded: false },
+      { id: 0, childrenExpanded: true },
+      { id: 1, childrenExpanded: true },
+      { id: 2, childrenExpanded: false },
     ]);
   });
 
@@ -69,11 +75,14 @@ describe('buildRowTree', () => {
     });
 
     expect(
-      Object.values(response.tree).map((node) => ({ id: node.id, expanded: node.expanded })),
+      Object.values(response.tree).map((node) => ({
+        id: node.id,
+        childrenExpanded: node.childrenExpanded,
+      })),
     ).to.deep.equal([
-      { id: 0, expanded: true },
-      { id: 1, expanded: true },
-      { id: 2, expanded: true },
+      { id: 0, childrenExpanded: true },
+      { id: 1, childrenExpanded: true },
+      { id: 2, childrenExpanded: true },
     ]);
   });
 
@@ -156,7 +165,7 @@ describe('buildRowTree', () => {
         0: {
           children: ['auto-generated-row-A-A'],
           depth: 0,
-          expanded: false,
+          childrenExpanded: false,
           groupingValue: 'A',
           id: 0,
           parent: null,
@@ -164,7 +173,7 @@ describe('buildRowTree', () => {
         'auto-generated-row-A-A': {
           children: [1],
           depth: 1,
-          expanded: false,
+          childrenExpanded: false,
           groupingValue: 'A',
           id: 'auto-generated-row-A-A',
           isAutoGenerated: true,
@@ -173,7 +182,7 @@ describe('buildRowTree', () => {
         1: {
           children: undefined,
           depth: 2,
-          expanded: false,
+          childrenExpanded: false,
           groupingValue: 'A',
           id: 1,
           parent: 'auto-generated-row-A-A',
