@@ -39,10 +39,10 @@ interface TempRowTreeNode extends Omit<GridRowTreeNodeConfig, 'children'> {
    ids: [0, 1, 2, 'auto-generated-row-B'],
    idRowsLookup: { 0: {...}, 1: {...}, 2: {...}, 'auto-generated-row-B': {} },
    tree: {
-     '0': { id: 0, parent: null, expanded: false, depth: 0, groupingValue: 'A' },
-     'auto-generated-row-B': { id: 'auto-generated-row-B', parent: null, expanded: false, depth: 0, groupingValue: 'B' },
-     '1': { id: 1, parent: 'auto-generated-row-B', expanded: false, depth: 1, groupingValue: 'A' },
-     '2': { id: 2, parent: 1, expanded: false, depth: 2, groupingValue: 'A' },
+     '0': { id: 0, parent: null, childrenExpanded: false, depth: 0, groupingValue: 'A' },
+     'auto-generated-row-B': { id: 'auto-generated-row-B', parent: null, childrenExpanded: false, depth: 0, groupingValue: 'B' },
+     '1': { id: 1, parent: 'auto-generated-row-B', childrenExpanded: false, depth: 1, groupingValue: 'A' },
+     '2': { id: 2, parent: 1, childrenExpanded: false, depth: 2, groupingValue: 'A' },
    },
    treeDepth: 3,
  }
@@ -67,7 +67,7 @@ export const buildRowTree = (params: GenerateRowTreeParams): GridRowGroupingResu
       const nodeName = row.path[depth];
       let nodeId: GridRowId;
 
-      const expanded =
+      const childrenExpanded =
         params.defaultGroupingExpansionDepth === -1 || params.defaultGroupingExpansionDepth > depth;
 
       let nodeNameConfig = nodeNameToIdSubTree[nodeName];
@@ -91,7 +91,7 @@ export const buildRowTree = (params: GenerateRowTreeParams): GridRowGroupingResu
           node = {
             id: nodeId,
             isAutoGenerated: true,
-            expanded,
+            childrenExpanded,
             parent: parentNode?.id ?? null,
             groupingValue: row.path[depth],
             depth,
@@ -104,7 +104,7 @@ export const buildRowTree = (params: GenerateRowTreeParams): GridRowGroupingResu
       } else {
         tempTree[row.id] = {
           id: row.id,
-          expanded,
+          childrenExpanded,
           parent: parentNode?.id ?? null,
           groupingValue: row.path[depth],
           depth,
