@@ -26,7 +26,7 @@ import { useGridSorting } from '../../_modules_/grid/hooks/features/sorting/useG
 import { useGridScroll } from '../../_modules_/grid/hooks/features/scroll/useGridScroll';
 import { useGridEvents } from '../../_modules_/grid/hooks/features/events/useGridEvents';
 import { useGridDimensions } from '../../_modules_/grid/hooks/features/dimensions/useGridDimensions';
-import { useGridTreeData } from '../../_modules_/grid/hooks/features/treeData';
+import { useGridTreeData } from '../../_modules_/grid/hooks/features/treeData/useGridTreeData';
 import { useGridRowGroupByColumns } from '../../_modules_/grid/hooks/features/rowGroupByColumns/useGridRowGroupByColumns';
 
 export const useDataGridProComponent = (apiRef: GridApiRef, props: GridComponentProps) => {
